refactor: :hammer: update strcuture
