+++ app/Http/Actions/ListUsersAction.php
@@ -2,24 +2,22 @@
 
 declare(strict_types=1);
 
-namespace App\Domains\User\Http\Actions;
+namespace App\Http\Actions;
 
-use App\Domains\User\Services\UserService;
+use App\Domain\User\Services\UserService;
 use App\Http\Responders\JsonResponder;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
-readonly class ShowUserAction
+readonly class ListUsersAction
 {
     public function __construct(private JsonResponder $responder, private UserService $userService)
     {
     }
 
-    public function __invoke(Request $request, Response $response, array $args): Response
+    public function __invoke(Request $request, Response $response): Response
     {
-        $userId = (int)$args['id'];
-
-        $user = $this->userService->getById($userId);
+        $user = $this->userService->getAll();
 
         return $this->responder->respond($response, $user);
     }
